Move ArtifactDeployWorkerTaskOutput as it was in the wrong package
