+++ impl/src/main/saker/maven/support/impl/deploy/ArtifactDeployWorkerTaskFactory.java
@@ -43,8 +43,8 @@ import saker.build.thirdparty.saker.util.ImmutableUtils;
 import saker.build.thirdparty.saker.util.io.SerialUtils;
 import saker.build.util.property.BuildTimeExecutionProperty;
 import saker.maven.support.api.ArtifactCoordinates;
-import saker.maven.support.api.ArtifactDeployWorkerTaskOutput;
 import saker.maven.support.api.MavenOperationConfiguration.RepositoryConfiguration;
+import saker.maven.support.api.deploy.ArtifactDeployWorkerTaskOutput;
 import saker.maven.support.impl.MavenImplUtils;
 import saker.maven.support.main.deploy.DeployArtifactsTaskFactory;
 import saker.maven.support.thirdparty.org.eclipse.aether.DefaultRepositorySystemSession;
